arch: rev 2 — address codex review findings (sbd#185)
Six codex findings from the rev 1 review addressed:

1. Missing option definitions for --app / --resource / etc. Added full
   Option def list to the design doc §3 (this commit only touches stubs).

2. Error exhaustiveness asserted, not designed. Design doc §5 now names
   the formatAndExit dispatcher with absurd() tail.

3. --as daemon-bypass not airtight: TransportOptions.daemonReachable
   was a boolean field (suggested socket probe before decision). Replaced
   with a lazy `probeDaemon?: () => Effect<boolean>` thunk invoked ONLY
   on the env-fallback branch. decideTransport now returns
   Effect<TransportDecision, never> so the probe is a side effect the
   graph records (Invariant §4.2).

4. ProfileRecord.registeredAt required broke back-compat. Made optional.
   loadLayeredConfig now documented as unknown-key tolerant.

5. 5-module budget: conversations-archival.ts was a 6th file. Deleted;
   three handler stubs inlined at the bottom of
   cli/commands/conversations.ts via barrel edit (architect rule single-
   file barrel carve-out). File count: 5 new modules.

6. Q-AS-1 invented --version default. Reclassified as ESCALATE to
   /safer:spec for rev 4 (Principle 8: no sidesteps).

Design doc rev 2: will be posted as reply comment on sbd#177.
